@@ -1,6 +1,6 @@
 {
   "name": "@angular/fire",
-  "version": "6.0.2",
+  "version": "6.0.3",
   "description": "The official library of Firebase and Angular.",
   "private": true,
   "scripts": {
@@ -64,7 +64,7 @@
     "tslib": "^1.10.0 || ^2.0.0",
     "ws": "^7.2.1",
     "xhr2": "^0.1.4",
-    "zone.js": "~0.10.2"
+    "zone.js": "0.10.2"
   },
   "optionalDependencies": {
     "bufferutil": "^4.0.1",
@@ -77,6 +77,7 @@
     "@angular/cli": "^9.0.0 || ^10.0.0",
     "@angular/compiler-cli": "^9.0.0 || ^10.0.0",
     "@angular/platform-server": "^9.0.0 || ^10.0.0",
+    "@firebase/app-types": "^0.6.1",
     "@types/fs-extra": "^7.0.0",
     "@types/gzip-size": "^5.1.1",
     "@types/inquirer": "^0.0.44",
@@ -84,35 +85,36 @@
     "@types/node": "^12.6.2",
     "@types/request": "0.0.30",
     "@types/semver": "^7.1.0",
-    "codelyzer": "^5.0.0",
+    "codelyzer": "^6.0.0",
     "concurrently": "^2.2.0",
     "conventional-changelog-cli": "^1.2.0",
-    "firebase-functions-test": "^0.2.0",
+    "firebase-functions-test": "^0.2.2",
     "gzip-size": "^5.1.1",
     "jasmine": "^3.4.0",
-    "jasmine-core": "^3.4.0",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "jasmine-core": "~3.5.0",
+    "karma": "~5.0.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "ncp": "^2.0.0",
-    "ng-packagr": "^9.0.0 || ^10.0.0",
+    "ng-packagr": "^10.1.0",
     "pretty-size": "^2.0.0",
-    "protractor": "3.0.0",
+    "protractor": "~7.0.0",
     "reflect-metadata": "0.1.2",
     "replace-in-file": "^5.0.2",
     "rimraf": "^2.5.4",
+    "rollup": "^2.22.0",
     "schematics-utilities": "^2.0.1",
     "shelljs": "^0.8.0",
     "systemjs": "^0.19.16",
     "systemjs-builder": "^0.15.7",
     "traceur": "0.0.96",
     "tsconfig-paths": "^3.8.0",
-    "tsickle": "^0.35.0",
-    "tslint": "^5.17.0",
+    "tsickle": "^0.39.1",
+    "tslint": "~6.1.0",
     "typedoc": "^0.16.4",
-    "typescript": "~3.6.4"
+    "typescript": "~3.9.7"
   },
   "typings": "index.d.ts"
 }
