@@ -9,7 +9,7 @@ export class AngularFirestoreModule {
   /**
    * Attempt to enable persistent storage, if possible
    */
-  static enablePersistence(persistenceSettings?: PersistenceSettings): ModuleWithProviders {
+  static enablePersistence(persistenceSettings?: PersistenceSettings): ModuleWithProviders<AngularFirestoreModule> {
     return {
       ngModule: AngularFirestoreModule,
       providers: [
