@@ -1,6 +1,7 @@
-export const environment = {
-  production: true,
-  firebase: {
+importScripts('https://www.gstatic.com/firebasejs/7.21.1/firebase-app.js');
+importScripts('https://www.gstatic.com/firebasejs/7.21.1/firebase-messaging.js');
+
+firebase.initializeApp({
     apiKey: 'AIzaSyA7CNE9aHbcSEbt9y03QReJ-Xr0nwKg7Yg',
     authDomain: 'aftest-94085.firebaseapp.com',
     databaseURL: 'https://aftest-94085.firebaseio.com',
@@ -9,6 +10,6 @@ export const environment = {
     messagingSenderId: '480362569154',
     appId: '1:480362569154:web:2fe6f75104cdfb82f50a5b',
     measurementId: 'G-CBRYER9PJR'
-  },
-  vapidKey: 'BIDPctnXHQDIjcOXxDS6qQcz-QTws7bL8v7UPgFnS1Ky5BZL3jS3-XXfxwRHmAUMOk7pXme7ttOBvVoIfX57PEo'
-};
+});
+
+const messaging = firebase.messaging();
